[clangd] IncludeCleaner: Respect main file forward declarations
Current behavior won't suggest unused includes, and will suggest
includes to add, even if there's a forward declaration in the
main file. This isn't intuitive behavior expected by users. So
if there's a forward declaration in the main file, consider that
the only usage needed, as expected.

This will create false positives when the forward declaration is
wrong and not sufficient for the usage, but IncludeCleaner does
not currently make any determination regarding validity of
forward declarations.